Rename MemberProfilesController to Account::MembersController.
