+++ app/controllers/member_profiles/passwords_controller.rb
@@ -10,7 +10,7 @@ class MemberProfiles::PasswordsController < ApplicationController
     @user = current_user
     if @user.update_with_password(user_params)
       bypass_sign_in(@user)
-      redirect_to member_profile_url, success: "Your password has been updated."
+      redirect_to account_member_url, success: "Your password has been updated."
     else
       redirect_to edit_member_profiles_password_url, warning: @user.errors.full_messages.join(". ").to_s
     end
