+++ app/views/account/members/_form.html.erb
@@ -1,4 +1,4 @@
-<%= form_with(model: member, url: member_profile_path, builder: SpectreFormBuilder) do |form| %>
+<%= form_with(model: member, url: account_member_path, builder: SpectreFormBuilder) do |form| %>
 
   <%= render 'admin/members/fields', form: form %>
