+++ app/views/account/members/show.html.erb
@@ -6,7 +6,7 @@
     </button>
     <br> <br>
     <button class="primary-btn btn-margin">
-      <%= link_to "Edit Member Profile", edit_member_profile_path %>
+      <%= link_to "Edit Member Profile", edit_account_member_path %>
     </button>
   </div>
 </div>
