+++ app/views/layouts/_user.html.erb
@@ -11,7 +11,7 @@
           <li class="menu-item"><%= link_to "Schedule an Appointment", new_account_appointment_path %></li>
           <li class="menu-item"><%= link_to "My Appointments", account_appointments_path %></li>
           <li class="menu-item"><%= link_to "View Loan History", member_loan_history_path %></li>
-          <li class="menu-item"><%= link_to "Account Settings", member_profile_path %></li>
+          <li class="menu-item"><%= link_to "Account Settings", account_member_path %></li>
           <li class="menu-item"><%= link_to('Logout', destroy_user_session_path, method: :delete) %></li>
         </ul>
       </div>
