+++ app/views/layouts/application.html.erb
@@ -60,7 +60,7 @@
                 <li class="nav-item"><%= link_to "Schedule an Appointment", new_account_appointment_path %></li>
                 <li class="nav-item"><%= link_to "My Appointments", account_appointments_path %></li>
                 <li class="nav-item"><%= link_to "View Loan History", member_loan_history_path %></li>
-                <li class="nav-item"><%= link_to "Account Settings", member_profile_path %></li>
+                <li class="nav-item"><%= link_to "Account Settings", account_member_path %></li>
               </ul>
                 <li class="nav-item logout-link"><%= link_to('Logout', destroy_user_session_path, method: :delete) %></li>
             <% else %>
