+++ config/routes.rb
@@ -22,6 +22,7 @@ Rails.application.routes.draw do
   namespace :account do
     resources :holds, only: [:create]
     resources :appointments, only: [:index, :new, :create]
+    resource :member, only: [:show, :edit, :update]
   end
 
   get "member/history", to: "members#history", as: "member_loan_history"
@@ -105,7 +106,6 @@ Rails.application.routes.draw do
 
   get "/s/:id", to: "short_links#show", as: :short_link
 
-  resource :member_profile, only: [:show, :edit, :update]
   namespace :member_profiles do
     resource :password, only: [:edit, :update]
   end
