--- test/controllers/member_profiles_controller_test.rb
@@ -1,32 +1,0 @@
-require "test_helper"
-
-class MemberProfilesControllerTest < ActionDispatch::IntegrationTest
-  include Devise::Test::IntegrationHelpers
-
-  setup do
-    @member = create(:verified_member_with_membership)
-    @user = create(:user, member: @member)
-    sign_in @user
-  end
-
-  test "member views profile" do
-    get member_profile_url
-    assert_response :success
-  end
-
-  test "member edits profile" do
-    get edit_member_profile_url
-    assert_response :success
-  end
-
-  test "member updates profile" do
-    patch member_profile_url, params: {member: {full_name: "Updated Name"}}
-    assert_redirected_to member_profile_url
-    assert_equal "Updated Name", @member.full_name
-  end
-
-  test "member update redirects to edit when provided invalid params" do
-    patch member_profile_url, params: {member: {full_name: nil}}
-    assert_template :edit
-  end
-end
